--- src/execute/Update.h
@@ -1,53 +1,0 @@
-//
-//  Copyright (c) 2014-2016 CASM Organization
-//  All rights reserved.
-//
-//  Developed by: Florian Hahn
-//                Philipp Paulweber
-//                Emmanuel Pescosta
-//                https://github.com/casm-lang/libcasm-fe
-//
-//  This file is part of libcasm-fe.
-//
-//  libcasm-fe is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-//
-//  libcasm-fe is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-//  GNU General Public License for more details.
-//
-//  You should have received a copy of the GNU General Public License
-//  along with libcasm-fe. If not, see <http://www.gnu.org/licenses/>.
-//
-
-#ifndef _LIB_CASMFE_UPDATE_H_
-#define _LIB_CASMFE_UPDATE_H_
-
-#include "../Value.h"
-#include "../various/location.hh"
-
-/**
- * @brief Represents an update
- */
-struct Update
-{
-    value_t value; /**< The value of the update */
-    const std::vector<value_t>* args; /**< The function arguments of the update */
-    const yy::location* location; /**< The source-code location of the update producer */
-    uint32_t func; /**< The function uid of the update */
-};
-
-#endif // _LIB_CASMFE_UPDATE_H_
-
-//
-//  Local variables:
-//  mode: c++
-//  indent-tabs-mode: nil
-//  c-basic-offset: 4
-//  tab-width: 4
-//  End:
-//  vim:noexpandtab:sw=4:ts=4:
-//
